+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -91,6 +91,12 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`cppcoreguidelines-no-temporary-expr-stmt
+  <clang-tidy/checks/cppcoreguidelines/no-temporary-expr-stmt>` check.
+
+  This implements ES.84: Don’t try to declare a local variable with no name.
+  Intended to catch errors due to missing a name for a scope guard variable.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^